+++ CHANGELOG.rst
@@ -37,6 +37,7 @@ In development.
 * Added a custom user model. This model behaves identically to the default user
   model, but we will be able to customize it in the future if the need arises.
 * Changed the default Python image to version 3.7 in GitLab CI.
+* Moved email-based error-reporting settings to ``settings/prod.py``.
 
 ----